rename array like
