+++ src/protobuf/input/arrayLike/index.ts
